llm_config_optimizer/requirements.txt
# LLM Configuration Optimizer Requirements

# Core dependencies
numpy>=1.20.0
scipy>=1.7.0

# LLM API
google-generativeai>=0.3.0

# Environment management
python-dotenv>=0.19.0

# Visualization (optional)
matplotlib>=3.3.0
